Removed resolve if IP is provided
hostname String parameter is parsed to byte array if possible and used to create an InetAddress instance. If parsing is not possible, the hostname (if not null) is considered to not be an IP and needs to be resolved before it can be used. Null will default the hostname to localhost